Added bot.py, config.py, utils/mysql_helper.py
Bot.py just contains the basic code to get the bot online and reading through the cogs folder.

Config.py just imports all the stuff from the dotenv file for me to use in the bot.

utils/mysql_helper.py is a custom mysql helper for ease of use using the python-mysql-connector library within the bot.