Updated bot.py. Added cogs/moderation.py
Changed on_ready event to check to see if the guilds table exists in the database and if not to create it when the bot starts up. Added on_guild_join event to insert the guild information into the database when the bot joins a new guild.

Added cogs/moderation.py contains basic helper functions for generating embeds that are going to be used, generating punishment ids and checking for mod log channels in the database to send logs to, it also includes a ban command mainly to test out the helper functions but the base of the command will stick.