Update cogs/moderation.py
Updated cogs/moderation.py, added 2 new helper functions for the timeout command, parse duration and format duration, these are used to parse the duration for the discord api and format it to put it into a nice embed. Also added the timeout command in the same format as the ban and kick command to keep the code consistent.