Added .env.example, requirements.txt. Updated README.md
Added .env.example to include an example configuration of the bots env file for self hosting.

Added requirements.txt for ease of installing librariess that the bot is using for self hosting.

Updated README.md to give an overview of the project, a roadmap and instructions on how to download and self host the bot.