Updated cogs/leveling.py
Updated cogs/leveling.py, added in the rank and leaderboard commands,  running the rank command will generate an image using the generate_level_card function and send it to the channel the command was ran in. The leaderboard command will get the top 10 people that have the highest level and xp and display them in an embed.